refactor(directory): :recycle: charts/preset/ -> Chart/
