+++ app/src/components/elements/Chart/CoalitionDynamicChart.tsx
