+++ app/src/pages/HomePage/contents/LastExamResult.tsx
@@ -1,6 +1,6 @@
 import { gql } from '@/__generated__';
 import { Spinner } from '@/components/common';
-import { BarChart } from '@/components/elements/charts/presets/BarChart';
+import { BarChart } from '@/components/elements/Chart';
 import { useQuery } from '@apollo/client';
 
 const GET_LAST_EXAM_RESULT = gql(/* GraphQL */ `
