+++ app/src/pages/TotalPage/contents/ProjectInfo.tsx
@@ -9,7 +9,7 @@ import {
   VStack,
 } from '@/components/common';
 import { Label } from '@/components/common/Label';
-import { PieChart } from '@/components/elements/charts/presets/PieChart';
+import { PieChart } from '@/components/elements/Chart';
 import { useSearchBar } from '@/components/elements/SearchBar/hooks/useSearchBar';
 import { ProjectSearchBar } from '@/components/elements/SearchBar/ProjectSearchBar';
 import { isEnterKeyReleased } from '@/utils/isEnterKeyReleased';
